comment rating add buffer in commentable model

+++ db/migrate/20151114055557_add_rating_to_driver.rb
@@ -1,6 +1,6 @@
 class AddRatingToDriver < ActiveRecord::Migration
   def change
-    add_column :drivers, :rating, :float
-    add_column :trips, :rating, :float
+    add_column :drivers, :rating, :decimal
+    add_column :trips, :rating, :decimal
   end
 end
